Split electron process in smaller modules
